Extract e2e open table helpers
