+++ frontend/test/__support__/e2e/cypress.js
@@ -9,33 +9,7 @@ export const version = require("../../../../version.json");
 export * from "./helpers/e2e-setup-helpers";
 export * from "./helpers/e2e-ui-elements-helpers";
 export * from "./helpers/e2e-dashboard-helpers";
-
-export function openTable({ database = 1, table, mode = null } = {}) {
-  const url = "/question/new?";
-  const params = new URLSearchParams({ database, table });
-
-  if (mode === "notebook") {
-    params.append("mode", mode);
-  }
-
-  cy.visit(url + params.toString());
-}
-
-export function openProductsTable({ mode } = {}) {
-  return openTable({ table: 1, mode });
-}
-
-export function openOrdersTable({ mode } = {}) {
-  return openTable({ table: 2, mode });
-}
-
-export function openPeopleTable({ mode } = {}) {
-  return openTable({ table: 3, mode });
-}
-
-export function openReviewsTable({ mode } = {}) {
-  return openTable({ table: 4, mode });
-}
+export * from "./helpers/e2e-open-table-helpers";
 
 export function setupLocalHostEmail() {
   // Email info
